refactor: move to Schedule module
