--- client/src/components/Schedule/index.ts
@@ -1,7 +1,0 @@
-export { TableView } from './TableView';
-export { SettingsPanel } from './SettingsPanel';
-export { ManageEventModalForm } from './ManageEventModalForm';
-export { TaskDetails, EventDetails } from './EventDetails';
-export { getTagStyle } from './utils';
-export { default as useScheduleSettings } from './useScheduleSettings';
-export type { ScheduleSettings } from './useScheduleSettings';
