+++ client/src/modules/Schedule/components/EventDetails/index.ts
