+++ client/src/modules/Schedule/components/SettingsPanel/AddEventButton.tsx
