+++ client/src/modules/Schedule/components/SettingsPanel/index.ts
