+++ client/src/modules/Schedule/components/TableView/index.ts
