+++ client/src/modules/Schedule/constants.ts
