+++ server/.env.example
@@ -7,14 +7,3 @@ RSSHCOOL_API_PG_DATABASE=rs_school
 
 RSSHCOOL_API_ADMIN_USERNAME=admin
 RSSHCOOL_API_ADMIN_PASSWORD=password
-
-# set to false to disable "admin" mode
-RSSHCOOL_DEV_ADMIN=true
-
-# mentors 2020 q1
-# RSSHCOOL_DEV_USERNAME=mikhama
-# RSSHCOOL_DEV_USERNAME=dmitryromaniuk
-
-# students 2020 q1
-# RSSHCOOL_DEV_USERNAME=yuliahope
-RSSHCOOL_DEV_USERNAME=yuliahope
